homework

+++ tareas/homework09/Programa2.java
@@ -6,8 +6,8 @@ class Programa2 {
     Scanner sc = new Scanner(System.in);
 
     int num1, num2;
-    num1 = sc.nextDouble();
-    num2 = sc.nextDouble();
+    num1 = sc.nextInt();
+    num2 = sc.nextInt();
 
     if(num1 > num2) {
       System.out.println(num2);
